+++ NEWS.md
@@ -1,4 +1,4 @@
-# arules 1.6-0.1 (xx/xx/2018)
+# arules 1.6-1 (04/04/2018)
 
 ## Bug Fixes
 * discretize now handles NAs in equal frequency (reported by yarik1988).
@@ -6,7 +6,7 @@
 
 ## New Features
 * rules and itemsets gained a method for nitems.
-
+  
 
 # arules 1.6-0 (2/28/2018)